jwt removal, file moving
